Test for Vxlan protocol.
target of this patch is to add a test for the vxlan protocol,
Vxlan is a network protocol support on Ofed and Mellanox upstream.
The test here creates a new qp - raw qp, for ethernet interfaces,
add to it rules for the steering in order to catch packet that match
the rule, in this way we check the hole subsystem, userspace and kernel,
we also send a wrong packet and check that it is not matching the
steering rules.

Issue: 766367
Signed-off-by: Moses Reuben <mosesr@mellanox.com>