Test for Flow tag .
target of this patch is to add a test for the Flow tag,
Flow tag is a network protocol supported on Ofed and Mellanox upstream.
The flow tag specification is used to identify flow with
specific tag-id.
The test here creates a raw qp, for ethernet interfaces,
add to it rules for the steering in order to catch packet that match
the rule, in this way we check the hole subsystem, userspace and kernel,
we also send a wrong packet and check that it is not matching the
steering rules.

Issue: 784340
Change-Id: I48ef772ff45cba1441b84825e87a9e2c05708734
Signed-off-by: Moses Reuben <mosesr@mellanox.com>